Moved Alias directive tests into UnsafeDeserializationAnalyzerTests class and added categories to the tests
